move unique serialization tests to kernel tests
