+++ tensorflow/python/data/experimental/kernel_tests/unique_test.py
@@ -21,6 +21,7 @@ from __future__ import print_function
 from absl.testing import parameterized
 
 from tensorflow.python.data.experimental.ops import unique
+from tensorflow.python.data.kernel_tests import checkpoint_test_base
 from tensorflow.python.data.kernel_tests import test_base
 from tensorflow.python.data.ops import dataset_ops
 from tensorflow.python.framework import combinations
@@ -93,6 +94,17 @@ class UniqueTest(test_base.DatasetTestBase, parameterized.TestCase):
         _ = dataset_ops.Dataset.from_generator(lambda: [],
                                                dtype).apply(unique.unique())
 
+class UniqueDatasetCheckpointTest(checkpoint_test_base.CheckpointTestBase,
+                                  parameterized.TestCase):
+
+  @combinations.generate(test_base.default_test_combinations())
+  def testUnique(self):
+
+    def build_dataset(num_elements, unique_elem_range):
+      return dataset_ops.Dataset.range(num_elements).map(
+          lambda x: x % unique_elem_range).apply(unique.unique())
+
+    self.run_core_tests(lambda: build_dataset(200, 100), 100)
 
 if __name__ == "__main__":
   test.main()
